rename: test 폴더의 schedule -> scheduler
