+++ backend/src/test/java/com/back/api/order/scheduler/OrderSchedulerFailTest.java
@@ -1,4 +1,4 @@
-package com.back.api.order.schedule;
+package com.back.api.order.scheduler;
 
 import com.back.domain.member.entity.Member;
 import com.back.domain.member.entity.Role;
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.TestPropertySource;
 
 import java.time.LocalDateTime;
